Fix description of -P option in wrapper manpage.
stunnel4 (and, by extention, the stunnel3 wrapper) does not have the special
handling of directories for the pid file. Also, to avoid creating one, the
correct way is to provide an empty arg, not 'none'. (Closes: #398012)

Aditionaly, change version number in changelog file to -5~1, to avoid having
a publicily available version number that conflicts with the next predicted
upload.