Revert the creation of a pidfile by stunnel3.
With the previous commit and this one, we now have the patch if we
should need it.  However, it seems to me that it would be kind of weird
to release a Debian package version that follows upstream in disabling
the automatic stunnel4 pidfile creation and, at the same time, adds a
Debian-specific patch to enable the automatic stunnel3 pidfile creation.