Rework the patches a bit.
Update the description of the 01-fix-paths patch.

Move the tools/script.sh modification from 01-fix-paths to
02-rename-binary and change "stunnel" to "stunnel4".

Drop the 08-client-example patch, it was actually applied upstream
and we added the same chunk of text a second time :/

Drop the 11-no-rle-compression patch, the OpenSSL bug has been fixed
somewhere in the 1.x release timeframe.

Discussed with:		upstream