Run the build in tools/, too.
For historical reasons, the files in the tools/ source subdirectory were
not regenerated at package build time.  The transition to debhelper
override rules kind of fixed that - since there was no override for only
running "make check" in src/ and doc/, it was also run in tools/, which
kind of inadvertently led to a "make all" run there :)  Of course, this
would FTBFS with the "nocheck" package build option.

Closes:		#804292
Reported by:	Daniel Schepler <dschepler@gmail.com>